refactor: create a new bitcoin module with bitcoin related stuff
