+++ signer/src/bitcoin/mod.rs
@@ -1,5 +1,9 @@
 //! Contains functionality for interacting with the Bitcoin blockchain
 
+pub mod fees;
+pub mod utxo;
+pub mod zmq;
+
 /// Represents the ability to interact with the bitcoin blockchain
 pub trait BitcoinInteract {
     /// Error type
