+++ signer/src/lib.rs
@@ -8,7 +8,6 @@ pub mod codec;
 pub mod config;
 pub mod ecdsa;
 pub mod error;
-pub mod fees;
 pub mod message;
 pub mod network;
 pub mod packaging;
@@ -18,9 +17,7 @@ pub mod storage;
 pub mod testing;
 pub mod transaction_coordinator;
 pub mod transaction_signer;
-pub mod utxo;
 pub mod wsts_state_machine;
-pub mod zmq;
 
 /// Package version
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
