+++ signer/tests/integration/rbf.rs
@@ -12,14 +12,14 @@ use bitcoincore_rpc::Error as BtcRpcError;
 use bitcoincore_rpc::RpcApi;
 use rand::distributions::Uniform;
 use rand::Rng;
-use signer::utxo::DepositRequest;
-use signer::utxo::Fees;
-use signer::utxo::RequestRef;
-use signer::utxo::SbtcRequests;
-use signer::utxo::SignerBtcState;
-use signer::utxo::SignerUtxo;
-use signer::utxo::UnsignedTransaction;
-use signer::utxo::WithdrawalRequest;
+use signer::bitcoin::utxo::DepositRequest;
+use signer::bitcoin::utxo::Fees;
+use signer::bitcoin::utxo::RequestRef;
+use signer::bitcoin::utxo::SbtcRequests;
+use signer::bitcoin::utxo::SignerBtcState;
+use signer::bitcoin::utxo::SignerUtxo;
+use signer::bitcoin::utxo::UnsignedTransaction;
+use signer::bitcoin::utxo::WithdrawalRequest;
 
 use crate::utxo_construction::generate_withdrawal;
 use crate::utxo_construction::make_deposit_request;
